chore: extract timestamp value to pg-srv
