+++ rust/cubesql/pg-srv/Cargo.toml
@@ -22,6 +22,7 @@ thiserror = "2"
 chrono = { version = "0.4", package = "chrono", default-features = false, features = [
     "clock",
 ], optional = true }
+chrono-tz = "0.8"
 
 [dev-dependencies]
 hex = "0.4.3"
